refactor: move all instruction logic to instruction.c
